requirements.txt
requests
BeautifulSoup
python-dotenv